requirements.txt
Flask==2.1.0
Gunicorn==20.1.0
mysqlclient==2.1.1
python-dotenv==0.21.0
validators==0.20.0
git+https://github.com/openai/whisper.git#egg=whisper
youtube_dl==2021.12.17